fix C* tests
