+++ cassandra/src/test/java/io/jaegertracing/spark/dependencies/cassandra/CassandraDependenciesJobTest.java
@@ -1,44 +1,150 @@
-package io.jaegertracing.spark.dependencies.cassandra;
+package io.jaegertracing.spark.dependencies;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
+import static org.junit.Assert.assertEquals;
 
-import static io.jaegertracing.spark.dependencies.cassandra.CassandraDependenciesJob.parseHosts;
-import static io.jaegertracing.spark.dependencies.cassandra.CassandraDependenciesJob.parsePort;
-
+import brave.Tracing;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.github.dockerjava.api.model.Link;
+import com.uber.jaeger.Tracer;
+import io.jaegertracing.spark.dependencies.cassandra.CassandraContainer;
+import io.jaegertracing.spark.dependencies.cassandra.CassandraDependenciesJob;
+import io.jaegertracing.spark.dependencies.common.DependenciesDerivator;
+import io.jaegertracing.spark.dependencies.common.JaegerTestDriverContainer;
+import io.jaegertracing.spark.dependencies.common.TracersGenerator;
+import io.jaegertracing.spark.dependencies.common.rest.DependencyLink;
+import io.jaegertracing.spark.dependencies.common.rest.JsonHelper;
+import io.jaegertracing.spark.dependencies.common.rest.RestResult;
+import io.jaegertracing.spark.dependencies.common.tree.Node;
+import io.jaegertracing.spark.dependencies.common.tree.TracingWrapper.JaegerWrapper;
+import io.jaegertracing.spark.dependencies.common.tree.TracingWrapper.ZipkinWrapper;
+import io.jaegertracing.spark.dependencies.common.tree.Traversals;
+import io.jaegertracing.spark.dependencies.common.tree.TreeGenerator;
+import java.io.IOException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
+import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.Response;
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Test;
+import org.testcontainers.containers.GenericContainer;
 
+/**
+ * @author Pavol Loffay
+ */
 public class CassandraDependenciesJobTest {
-  @Test
-  public void parseHosts_ignoresPortSection() {
-    assertThat(parseHosts("1.1.1.1:9142"))
-        .isEqualTo("1.1.1.1");
-  }
 
-  @Test
-  public void parseHosts_commaDelimits() {
-    assertThat(parseHosts("1.1.1.1:9143,2.2.2.2:9143"))
-        .isEqualTo("1.1.1.1,2.2.2.2");
-  }
+    private static OkHttpClient okHttpClient = new OkHttpClient.Builder()
+            .build();
 
-  @Test
-  public void parsePort_ignoresHostSection() {
-    assertThat(parsePort("1.1.1.1:9142"))
-        .isEqualTo("9142");
-  }
+    private ObjectMapper objectMapper = JsonHelper.configure(new ObjectMapper());
 
-  @Test
-  public void parsePort_multiple_consistent() {
-    assertThat(parsePort("1.1.1.1:9143,2.2.2.2:9143"))
-        .isEqualTo("9143");
-  }
+    protected static int cassandraPort;
+    protected static String queryUrl;
+    protected static String collectorUrl;
+    protected static String zipkinCollectorUrl;
 
-  @Test
-  public void parsePort_defaultsTo9042() {
-    assertThat(parsePort("1.1.1.1"))
-        .isEqualTo("9042");
-  }
+    protected GenericContainer cassandra;
+    protected GenericContainer jaegerTestDriver;
 
-  @Test
-  public void parsePort_defaultsTo9042_multi() {
-  }
+    @Before
+    public void before() throws TimeoutException {
+        cassandra = new CassandraContainer("cassandra:3.9")
+                        .withExposedPorts(9042);
+        cassandra.start();
+        cassandraPort = cassandra.getMappedPort(9042);
+
+        jaegerTestDriver = new JaegerTestDriverContainer("jaegertracing/test-driver:latest")
+                        .withCreateContainerCmdModifier(cmd -> {
+                            cmd.withLinks(new Link(cassandra.getContainerId(), "cassandra"));
+                            cmd.withHostName("test_driver");
+                        })
+                    .withExposedPorts(14268, 16686, 8080, 9411);
+        jaegerTestDriver.start();
+        queryUrl = String.format("http://localhost:%d", jaegerTestDriver.getMappedPort(16686));
+        collectorUrl = String.format("http://localhost:%d", jaegerTestDriver.getMappedPort(14268));
+        zipkinCollectorUrl = String.format("http://localhost:%d", jaegerTestDriver.getMappedPort(9411));
+    }
+
+    @After
+    public void after() {
+        cassandra.stop();
+        jaegerTestDriver.stop();
+    }
+
+
+    @Test
+    public void testJaeger() throws IOException {
+        TreeGenerator<Tracer> treeGenerator = new TreeGenerator(
+            TracersGenerator.generateJaeger(2, collectorUrl));
+        Node<JaegerWrapper> root = treeGenerator.generateTree(15, 3);
+        Traversals.inorder(root, (node, parent) -> node.getTracingWrapper().get().getSpan().finish());
+        treeGenerator.getTracers().forEach(tracer -> tracer.getTracer().close());
+        waitRestTracesContains(root.getServiceName(), root.getTracingWrapper().operationName());
+
+        CassandraDependenciesJob.builder()
+                .logInitializer(LogInitializer.create("INFO"))
+                .contactPoints("localhost:" + cassandraPort)
+                .day(System.currentTimeMillis())
+                .keyspace("jaeger")
+                .build()
+                .run();
+
+        Request request = new Request.Builder()
+                .url(queryUrl + "/api/dependencies?endTs=" + System.currentTimeMillis())
+                .get()
+                .build();
+        DependenciesDerivator.serviceDependencies(root);
+        try (Response response = okHttpClient.newCall(request).execute()) {
+            assertEquals(200, response.code());
+            RestResult<DependencyLink> restResult = objectMapper.readValue(response.body().string(), new TypeReference<RestResult<DependencyLink>>() {});
+            assertEquals(null, restResult.getErrors());
+            assertEquals(DependenciesDerivator.serviceDependencies(root), DependenciesDerivator.serviceDependencies(restResult.getData()));
+        }
+    }
+
+    @Test
+    public void testZipkin() throws IOException {
+        TreeGenerator<Tracing> treeGenerator = new TreeGenerator(TracersGenerator.generateZipkin(2, zipkinCollectorUrl));
+        Node<ZipkinWrapper> root = treeGenerator.generateTree(15, 3);
+        Traversals.inorder(root, (node, parent) -> node.getTracingWrapper().get().getSpan().finish());
+        treeGenerator.getTracers().forEach(tracer -> tracer.getTracer().close());
+
+        waitRestTracesContains(root.getServiceName(), root.getTracingWrapper().operationName());
+
+        CassandraDependenciesJob.builder()
+            .logInitializer(LogInitializer.create("INFO"))
+            .contactPoints("localhost:" + cassandraPort)
+            .day(System.currentTimeMillis())
+            .keyspace("jaeger")
+            .build()
+            .run();
+
+        Request request = new Request.Builder()
+            .url(queryUrl + "/api/dependencies?endTs=" + System.currentTimeMillis())
+            .get()
+            .build();
+        DependenciesDerivator.serviceDependencies(root);
+        try (Response response = okHttpClient.newCall(request).execute()) {
+            assertEquals(200, response.code());
+            RestResult<DependencyLink> restResult = objectMapper.readValue(response.body().string(), new TypeReference<RestResult<DependencyLink>>() {});
+            assertEquals(null, restResult.getErrors());
+            assertEquals(DependenciesDerivator.serviceDependencies(root), DependenciesDerivator.serviceDependencies(restResult.getData()));
+        }
+    }
+
+    public void waitRestTracesContains(String service, String spanContainsThis) {
+        Request request2 = new Request.Builder()
+            .url(String.format("%s/api/traces?service=%s", queryUrl, service))
+            .get()
+            .build();
+        await().atMost(1, TimeUnit.MINUTES).until(() -> {
+            Response response = okHttpClient.newCall(request2).execute();
+            String body = response.body().string();
+            return body.contains(spanContainsThis);
+        });
+    }
 }
