+++ elasticsearch/pom.xml
@@ -23,5 +23,13 @@
       <artifactId>elasticsearch-spark-20_${scala.binary.version}</artifactId>
       <version>6.0.0-beta2</version>
     </dependency>
+
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>jaeger-spark-dependencies-common</artifactId>
+      <version>${project.version}</version>
+      <type>test-jar</type>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
