+++ jaeger-spark-dependencies-common/src/main/java/io/jaegertracing/spark/dependencies/LogInitializer.java
