+++ jaeger-spark-dependencies-common/src/test/java/io/jaegertracing/spark/dependencies/common/rest/DependencyLink.java
@@ -1,4 +1,4 @@
-package io.jaegertracing.spark.dependencies.rest;
+package io.jaegertracing.spark.dependencies.common.rest;
 
 /**
  * @author Pavol Loffay
