+++ jaeger-spark-dependencies-common/src/test/java/io/jaegertracing/spark/dependencies/common/rest/JsonHelper.java
@@ -1,4 +1,4 @@
-package io.jaegertracing.spark.dependencies.rest;
+package io.jaegertracing.spark.dependencies.common.rest;
 
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
