+++ jaeger-spark-dependencies-common/src/test/java/io/jaegertracing/spark/dependencies/common/tree/TracingWrapper.java
@@ -1,4 +1,4 @@
-package io.jaegertracing.spark.dependencies.tree;
+package io.jaegertracing.spark.dependencies.common.tree;
 
 import brave.Span.Kind;
 import brave.Tracing;
